+++ udata/core/organization/forms.py
@@ -3,7 +3,8 @@ from udata.forms import Form, ModelForm, fields, validators
 from udata.i18n import lazy_gettext as _
 
 from .models import (
-    Organization, MembershipRequest, Member, LOGO_SIZES, ORG_ROLES
+    Organization, MembershipRequest, Member, LOGO_SIZES, ORG_ROLES,
+    TITLE_SIZE_LIMIT, DESCRIPTION_SIZE_LIMIT
 )
 
 __all__ = (
@@ -17,11 +18,11 @@ __all__ = (
 class OrganizationForm(ModelForm):
     model_class = Organization
 
-    name = fields.StringField(_('Name'), [validators.DataRequired(), validators.Length(max=350)])
+    name = fields.StringField(_('Name'), [validators.DataRequired(), validators.Length(max=TITLE_SIZE_LIMIT)])
     acronym = fields.StringField(
         _('Acronym'), description=_('Shorter organization name'))
     description = fields.MarkdownField(
-        _('Description'), [validators.DataRequired(), validators.Length(max=20000)],
+        _('Description'), [validators.DataRequired(), validators.Length(max=DESCRIPTION_SIZE_LIMIT)],
         description=_('The details about your organization'))
     url = fields.URLField(
         _('Website'), description=_('The organization website URL'))